--- a/tests/model_regress/test_state.py
+++ b/tests/model_regress/test_state.py
@@ -1,0 +1,8 @@
+from django.db.models.base import ModelState, ModelStateFieldsCacheDescriptor
+from django.test import SimpleTestCase
+
+
+class ModelStateTests(SimpleTestCase):
+
+    def test_fields_cache_descriptor(self):
+        self.assertIsInstance(ModelState.fields_cache, ModelStateFieldsCacheDescriptor)