@charettes 2017-08-07
Neat trick but this optimization which I assume is mostly due to the `__init__()` removal won't be possible once #6721's is merged because it introduces `ModelState.fields_cache`. Might be worth defining `db` and `adding` at the class level anyway.