@sir-sigurd 2017-08-07
@charettes 
I think I could make `ModelState.fields_cache` cached property in this case.