@charettes 2017-08-07
@sir-sigurd that could work but it's hard to tell whether or not it will make the `select_related()` case, which will access `fields_cache` at initialization, so much worse performance wise that the optimization is not worth it for the cases that it isn't used.