@sir-sigurd 2017-08-07
@charettes this should be checked of course, but I think it could make faster even `select_related()` case, because there will be just one `fields_cache` initialization but two `Model` initialization per row.