@charettes 2017-08-07
@sir-sigurd you're probably right, let's wait until #6721 is merged to check it out. By the way, I was wondering if you knew about [djangobench](https://github.com/django/djangobench). I think it would need a bit of adjustments to get it running against master but it has a few benchmarks related to model initialization from what I remember.