@sir-sigurd 2017-08-10
@charettes I know about djangobench but that part
> it would need a bit of adjustments

deterred me from using it.

I updated PR, but had to use specific descriptor instead of `cached_property` for `fields_cache` because `cached_property` turned out to be too slow for this case (performance of `select_related` was decreased when all related objects are `None`).