@timgraham 2017-08-11
By the way, I fixed djangobench to work with django master in case you want to use it for this or any other patches.